omz: Disable compfix when running under pseudo
compfix checks the permissions of completion directories to belong to
the current user and not be group/world-writable. This fails when
running under fakeroot-psudo, as the completion dirs are still owned by
the outside user but compfix thinks it is root.
To keep the security risk small, disable compfix only when pseudo is
active.